% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_targets.R
\name{check_targets}
\alias{check_targets}
\title{Check target columns for agreement
Wrapper function for `check_row`, which checks three elements of a named vector against each other and returns TRUE (if they match) or FALSE (if they don't) while ignoring NAs.}
\usage{
check_targets(
  dataframe,
  cols = c("lut_Species_MitoFish", "BLAST_scientific_name", "Species")
)
}
\arguments{
\item{dataframe}{data frame object to check}

\item{cols}{specific columns in the \code{dataframe} to check}
}
\value{
a logical vector
}
\description{
check_targets
}
\author{
Matt Espe, Myfanwy Johnston
}
