% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaphos-package.R
\docType{package}
\name{elaphos-package}
\alias{elaphos-package}
\alias{elaphos}
\title{The 'elaphos' package}
\description{
Data and convenience functions for eDNA experiments and sampling conducted by Cramer Fish Sciences.
}
