% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_utils.R
\name{cq_to_lnconc}
\alias{cq_to_lnconc}
\alias{lnconc_to_cq}
\title{Convert CQ value to [eDNA] and vice-versa}
\usage{
cq_to_lnconc(Cq_values, std_curve_alpha, std_curve_beta)

lnconc_to_cq(x, std_curve_alpha, std_curve_beta, censor = TRUE, upper_Cq = 40)
}
\arguments{
\item{Cq_values}{numeric vector, value of CQ}

\item{std_curve_alpha}{the alpha (intercept) value for the
standard curve}

\item{std_curve_beta}{the beta (slope) value for the standard
curve}

\item{x}{log[eDNA] values}

\item{censor}{logical, whether Cq values larger than the upper_Cq
should be censored. If TRUE, these any values above the
threshold will be replaced with the upper_Cq value.}

\item{upper_Cq}{the max Cq value}
}
\value{
numeric vector, [eDNA] values

vector of Cq values
}
\description{
Convenience function to convert from Cq to [eDNA] and vice-versa
}
\details{
Mirror of the same functions from the artemis package (fishsciences.github.io/artemis)
}
\author{
Matt Espe
}
