% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report_table.R
\name{make_detection_table}
\alias{make_detection_table}
\title{Make Detection Data.frame}
\usage{
make_detection_table(
  df,
  threshold = 40,
  labs = c("ND", "(+)"),
  cq_col = "Cq",
  target_col = "Target",
  filter_col = "FilterID"
)
}
\arguments{
\item{df}{data.frame of CQ results, as returned by \code{read_qpcr}}

\item{threshold}{numeric, the threshold CQ, below which is considered a detection}

\item{labs}{character vector of length 2 corresponding to the
string to label for 0 (nondetect) and 1 (detect)}

\item{cq_col}{string, the name of the CQ column}

\item{target_col}{string, the name of the target column}

\item{filter_col}{string, the name of the filter column}
}
\value{
data.frame of detection table
}
\description{
Creates a detection data.frame from CQ results, suitable for summary reports.
}
\author{
Matt Espe
}
