% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qpcr.R
\name{read_qpcr}
\alias{read_qpcr}
\title{read_qpcr: a function to read the .xls files exported by QuantStudio}
\usage{
read_qpcr(
  File,
  Cq_threshold = 40,
  Experiment,
  Keep_controls = FALSE,
  Column_numbers = c(2, 4, 5, 9),
  Column_names = c("WellPosition", "FilterID", "Target", "Cq"),
  Control_strings = c("NTC", "CT+", "0", "BHOO", "DS", "BHOO+", "DS+")
)
}
\arguments{
\item{Cq_threshold}{the Cq value at which to assume non-detection}

\item{Experiment}{the name of the experiment or project associated with the qpcr data}

\item{file}{the path to the .xls file}
}
\value{
a dataframe with the following columns: "Experiment", "FilterID", "TechRep", "Target", "Cq", "TechRepID", "PlateID", "WellID"
}
\description{
this takes a single .xls file as its input.  It returns a dataframe.
}
