% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qpcr.R
\name{read_qpcr}
\alias{read_qpcr}
\title{read_qpcr: a function to read the .xls files exported by QuantStudio}
\usage{
read_qpcr(
  File,
  Cq_threshold = 40,
  Experiment,
  Keep_controls = FALSE,
  Column_numbers = c(2, 4, 5, 9),
  Column_names = c("WellPosition", "FilterID", "Target", "Cq"),
  Control_strings = c("NTC", "CT+", "^0$", "BHOO", "DS", "BHOO+", "DS+")
)
}
\arguments{
\item{File}{the path to the .xls file}

\item{Cq_threshold}{the Cq value at which to assume non-detection}

\item{Experiment}{the name of the experiment or project associated with the qpcr data}

\item{Keep_controls}{logical indicating whether you want to retain the FilterIDs with matching strings provided in the \code{Control_strings} argument}

\item{Column_numbers}{integer vector corresponding to the columns you want to keep from the raw qpcr output}

\item{Column_names}{character vector of column names to apply to the \code{Column_numbers}}

\item{Control_strings}{character vector of strings to check for in the FilterID column that indicate a control sample}
}
\value{
a data frame with the following columns: "Experiment", "TechRep", "PlateID", "Control", and those specified in Column_names.
}
\description{
this takes a single .xls file as its input.  It returns a dataframe.
}
